refactor(dev): isolate tilt and relocate tests
