+++ .dockerignore
@@ -1,11 +1,20 @@
 node_modules
 .git
-data
-server/dist
-server/data
-frontend/dist
 scratchpad
 *.db
 *.db-wal
 *.db-shm
 *.db-journal
+
+# Runtime/dev state
+.local
+.dev
+
+# Build outputs
+apps/*/dist
+apps/*/*.tsbuildinfo
+packages/*/dist
+packages/*/*.tsbuildinfo
+
+# Repo-local runtime data (legacy; new defaults live outside the repo)
+data
