+++ .gitignore
@@ -4,6 +4,7 @@ node_modules/
 # Build outputs
 dist/
 build/
+**/dist/
 *.tsbuildinfo
 
 # Environment
@@ -13,6 +14,7 @@ build/
 
 # Data
 /data/
+.local/
 *.db
 *.sqlite
 
@@ -34,7 +36,8 @@ Thumbs.db
 # Testing
 coverage/
 
-server/data/
+apps/gateway/data/
+packages/*/dist/
 
 # Misc
 *.bak
