+++ apps/gateway/test/api/settings.test.ts
