+++ infra/docker/Dockerfile.web.dev
@@ -1,24 +1,24 @@
-# Dev-mode web app: tsx watch (Gateway/Express) + vite dev server (React)
-# Used by Tilt with live_update for sub-second file syncs.
-# IMPORTANT: agent-service runs in its own deployment; do not start it here.
+# Dev-mode web stack: gateway (tsx watch) + package watchers + Vite.
+# Used by Tilt with live_update for scoped syncs.
 
 FROM node:22-alpine
 
 WORKDIR /app
 
-# Install deps (including devDependencies)
-COPY package.json package-lock.json ./
-COPY server/package.json ./server/
-COPY frontend/package.json ./frontend/
+COPY package.json package-lock.json tsconfig.base.json tsconfig.json ./
+COPY apps/gateway/package.json ./apps/gateway/package.json
+COPY apps/frontend/package.json ./apps/frontend/package.json
+COPY packages/contracts/package.json ./packages/contracts/package.json
+COPY packages/config/package.json ./packages/config/package.json
+COPY packages/data/package.json ./packages/data/package.json
+COPY packages/runtime/package.json ./packages/runtime/package.json
 RUN npm install
 
-# Copy source (overridden by Tilt live_update on changes)
-COPY . .
+COPY apps/gateway ./apps/gateway
+COPY apps/frontend ./apps/frontend
+COPY packages ./packages
+COPY scripts ./scripts
 
-# Gateway (3000) + Vite (5173)
 EXPOSE 3000 5173
 
-# Run only gateway + frontend here.
-# --host 0.0.0.0 on Vite is critical — without it, Vite only binds
-# localhost inside the container and is unreachable from outside.
 CMD ["npm", "run", "dev:web"]
