+++ ops/headlamp/scripts/generate-headlamp-token.sh
@@ -3,7 +3,8 @@ set -euo pipefail
 
 NAMESPACE="goldilocks"
 SERVICE_ACCOUNT="headlamp-admin"
-OUT_FILE=".dev/headlamp-token.txt"
+STATE_ROOT="${GOLDILOCKS_STATE_DIR:-${XDG_STATE_HOME:-$HOME/.local/state}/goldilocks}"
+OUT_FILE="${STATE_ROOT}/headlamp/headlamp-token.txt"
 TMP_FILE="${OUT_FILE}.tmp"
 
 mkdir -p "$(dirname "$OUT_FILE")"
